Extract main tab bar on launch actions into a separate file
